requirements.txt
Django==5.2.6
djangorestframework==3.15.2
channels==4.1.0
channels-redis==4.2.0
django-cors-headers==4.3.1
Pillow==10.4.0
redis==5.0.1
requests==2.31.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.10
daphne==4.0.0
